backend/sales_dashboard/requirements.txt
Package                       Version               Editable project location
----------------------------- --------------------- ---------------------------
anyio                         3.6.1
argon2-cffi                   21.3.0
argon2-cffi-bindings          21.2.0
asgiref                       3.8.1
asttokens                     2.0.5
attrs                         21.4.0
Babel                         2.10.1
backcall                      0.2.0
beautifulsoup4                4.11.1
bleach                        5.0.0
blinker                       1.6.3
certifi                       2022.5.18.1
cffi                          1.15.0
charset-normalizer            2.0.12
click                         8.1.7
colorama                      0.4.4
debugpy                       1.6.0
decorator                     5.1.1
defusedxml                    0.7.1
distlib                       0.3.7
Django                        4.2.dev20220728095248 c:\users\sreeraj v s\django
django-cors-headers           4.3.1
djangorestframework           3.15.1
djangorestframework-simplejwt 5.3.1
entrypoints                   0.4
executing                     0.8.3
fastjsonschema                2.15.3
filelock                      3.12.4
Flask                         3.0.0
Flask-Cors                    4.0.1
idna                          3.3
ipykernel                     6.13.1
ipython                       8.4.0
ipython-genutils              0.2.0
itsdangerous                  2.1.2
jedi                          0.18.1
Jinja2                        3.1.2
json5                         0.9.8
jsonschema                    4.6.0
jupyter-client                7.3.3
jupyter-core                  4.10.0
jupyter-server                1.17.1
jupyterlab                    3.4.3
jupyterlab-pygments           0.2.2
jupyterlab-server             2.14.0
MarkupSafe                    2.1.1
matplotlib-inline             0.1.3
mistune                       0.8.4
nbclassic                     0.3.7
nbclient                      0.6.4
nbconvert                     6.5.0
nbformat                      5.4.0
nest-asyncio                  1.5.5
notebook                      6.4.12
notebook-shim                 0.1.0
numpy                         1.26.4
packaging                     21.3
pandas                        2.2.2
pandocfilters                 1.5.0
parso                         0.8.3
pickleshare                   0.7.5
pip                           22.0.4
platformdirs                  3.11.0
prometheus-client             0.14.1
prompt-toolkit                3.0.29
psutil                        5.9.1
psycopg2                      2.9.3
psycopg2-binary               2.9.9
pure-eval                     0.2.2
pycparser                     2.21
Pygments                      2.12.0
PyJWT                         2.8.0
pyparsing                     3.0.9
pyrsistent                    0.18.1
python-dateutil               2.8.2
pytz                          2022.1
pywin32                       304
pywinpty                      2.0.5
pyzmq                         23.1.0
requests                      2.27.1
Send2Trash                    1.8.0
setuptools                    58.1.0
six                           1.16.0
sniffio                       1.2.0
soupsieve                     2.3.2.post1
sqlparse                      0.4.2
stack-data                    0.2.0
terminado                     0.15.0
tinycss2                      1.1.1
tornado                       6.1
traitlets                     5.2.2.post1
typing_extensions             4.12.1
tzdata                        2024.1
urllib3                       1.26.9
virtualenv                    20.24.5
wcwidth                       0.2.5
webencodings                  0.5.1
websocket-client              1.3.2
Werkzeug                      3.0.0
